requirements.txt
pandas>=1.5.3
numpy>=1.24.3
scikit-learn>=1.2.2
matplotlib>=3.7.1
seaborn>=0.12.2
joblib>=1.2.0
xgboost>=1.7.6
ipykernel>=6.22.0
statsmodels>=0.14.0
pymongo>=4.3.3
python-dotenv>=1.0.0
certifi>=2023.7.22
dill>=0.3.6
pyyaml>=6.0
imbalanced-learn>=0.10.1
scipy>=1.10.1
catboost>=1.1
fastapi>=0.95.2
uvicorn>=0.22.0
python-box>=5.4.1
lightgbm
boto3
mypy-boto3-s3
botocore
pydantic
mlflow
dagshub
dvc
# apache-airflow==2.8.3

# -e .